applications/wellness_companion/requirements.txt
fastapi
uvicorn
pydantic
openai==0.28.0
sentence-transformers
faiss-cpu
pandas
numpy